requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-on-heroku==1.1.2
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
mercadopago==2.1.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytest-mock==3.7.0
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
whitenoise==6.0.0
